+++ src/lib.rs
@@ -39,7 +39,6 @@ use logger::{Data as LoggerData, LOGGER};
 pub use logger::{Logger, Message};
 pub use object::Map;
 use object::Object;
-use once_cell::sync::Lazy;
 use options::Ownership;
 pub use options::{Options, Shutdown};
 pub use panic::set_hook;
@@ -48,7 +47,6 @@ use std::{
     fmt::{Display, Formatter, Result as FmtResult},
     os::raw::c_char,
     ptr,
-    sync::{Mutex, MutexGuard},
 };
 use thiserror::Error;
 use transport::State as TransportState;
@@ -171,15 +169,6 @@ impl Consent {
             sys::UserConsent::Given => Self::Given,
         }
     }
-}
-
-/// Global lock to prevent thread-safety issues with certain functions. This
-/// will hopefully be fixed upstream.
-static GLOBAL_LOCK: Lazy<Mutex<()>> = Lazy::new(|| Mutex::new(()));
-
-/// Convenience function to get a write lock on `GLOBAL_LOCK`.
-fn global_lock() -> MutexGuard<'static, ()> {
-    GLOBAL_LOCK.lock().expect("global lock poisoned")
 }
 
 /// Shuts down the Sentry client and forces transports to flush out.
@@ -347,7 +336,6 @@ pub fn user_consent() -> Consent {
 /// remove_user();
 /// ```
 pub fn remove_user() {
-    let _lock = global_lock();
     unsafe { sys::remove_user() }
 }
 
@@ -362,10 +350,7 @@ pub fn set_tag<S1: Into<String>, S2: Into<String>>(key: S1, value: S2) {
     let key = key.into().into_cstring();
     let value = value.into().into_cstring();
 
-    {
-        let _lock = global_lock();
-        unsafe { sys::set_tag(key.as_ptr(), value.as_ptr()) }
-    }
+    unsafe { sys::set_tag(key.as_ptr(), value.as_ptr()) }
 }
 
 /// Removes the tag with the specified `key`.
@@ -378,11 +363,7 @@ pub fn set_tag<S1: Into<String>, S2: Into<String>>(key: S1, value: S2) {
 /// ```
 pub fn remove_tag<S: Into<String>>(key: S) {
     let key = key.into().into_cstring();
-
-    {
-        let _lock = global_lock();
-        unsafe { sys::remove_tag(key.as_ptr()) }
-    }
+    unsafe { sys::remove_tag(key.as_ptr()) }
 }
 
 /// Sets extra information.
@@ -396,10 +377,7 @@ pub fn set_extra<S: Into<String>, V: Into<Value>>(key: S, value: V) {
     let key = key.into().into_cstring();
     let value = value.into().into_raw();
 
-    {
-        let _lock = global_lock();
-        unsafe { sys::set_extra(key.as_ptr(), value) }
-    }
+    unsafe { sys::set_extra(key.as_ptr(), value) }
 }
 
 /// Removes the extra with the specified `key`.
@@ -412,11 +390,7 @@ pub fn set_extra<S: Into<String>, V: Into<Value>>(key: S, value: V) {
 /// ```
 pub fn remove_extra<S: Into<String>>(key: S) {
     let key = key.into().into_cstring();
-
-    {
-        let _lock = global_lock();
-        unsafe { sys::remove_extra(key.as_ptr()) }
-    }
+    unsafe { sys::remove_extra(key.as_ptr()) }
 }
 
 /// Sets a context object.
@@ -430,10 +404,7 @@ pub fn set_context<S: Into<String>, M: Map + Into<Value>>(key: S, value: M) {
     let key = key.into().into_cstring();
     let value = value.into().into_raw();
 
-    {
-        let _lock = global_lock();
-        unsafe { sys::set_context(key.as_ptr(), value) }
-    }
+    unsafe { sys::set_context(key.as_ptr(), value) }
 }
 
 /// Removes the context object with the specified key.
@@ -446,11 +417,7 @@ pub fn set_context<S: Into<String>, M: Map + Into<Value>>(key: S, value: M) {
 /// ```
 pub fn remove_context<S: Into<String>>(key: S) {
     let key = key.into().into_cstring();
-
-    {
-        let _lock = global_lock();
-        unsafe { sys::remove_context(key.as_ptr()) }
-    }
+    unsafe { sys::remove_context(key.as_ptr()) }
 }
 
 /// Sets the event fingerprint.