+++ CHANGELOG.md
@@ -6,6 +6,26 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added `reinstall_backend`.
+- Enabled Breakpad support for MacOS.
+
+### Changed
+
+- Updated `sentry-native` to 8a02abf.
+- Removed minimum supported Windows SDK requirement.
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
 ## [0.2.0] - 2021-01-19
 
 ### Added